Task 363: rename orchestrator to router
