+++ services/pom.xml
@@ -20,7 +20,7 @@
         <module>pbchange</module>
         <module>pbhealth</module>
         <module>scopechange</module>
-        <module>orchestrator</module>
+        <module>router</module>
     </modules>
     <dependencyManagement>
         <dependencies>
