+++ services/router/pom.xml
@@ -7,13 +7,13 @@
         <groupId>org.boston</groupId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>orchestrator</artifactId>
+    <artifactId>router</artifactId>
     <modelVersion>4.0.0</modelVersion>
 
     <packaging>jar</packaging>
 
     <build>
-        <finalName>orchestrator</finalName>
+        <finalName>router</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -22,7 +22,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>orchestrator.Main</mainClass>
+                            <mainClass>router.Main</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
