+++ services/router/src/main/java/router/Env.java
@@ -1,4 +1,4 @@
-package orchestrator;
+package router;
 
 public class Env {
     /**
