+++ services/router/src/main/java/router/Router.java
@@ -1,16 +1,16 @@
-package orchestrator;
+package router;
 
 import org.apache.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import orchestrator.routes.Routes;
+import router.routes.Routes;
 
 import static spark.Spark.get;
 import static spark.Spark.port;
 
-public class Orchestrator {
-    private final static Logger logger = LoggerFactory.getLogger(Orchestrator.class);
+public class Router {
+    private final static Logger logger = LoggerFactory.getLogger(Router.class);
     public static void start() {
         port(Env.getPort()); // TODO: figure out how to make railway listen to this port (or how to get railway's port here)
         logger.info("Starting server");
