+++ services/router/src/main/java/router/routes/RouteAPI.java
@@ -1,4 +1,4 @@
-package orchestrator.routes;
+package router.routes;
 
 import bostonhttp.api.APIWrapperBase;
 import bostonhttp.api.APIWrapperBehaviors;
